Set `crossScalaVersions := Nil` in root project
Otherwise when 2.12.13 is selected and we try to run `update` in the
testSuite project, which is the only project that removes Scala 2.12.13
from its dependencies, it runs instead with Scala 2.13.5 and then it fails
to fetch some artifacts since the locales plugin is generating the artifacts
for 2.12 instead.
As suggested in the sbt issues tracker, the solution is to set
`scalaCrossVersions` to `Nil` for the aggregated project.